@@ -4,8 +4,8 @@ version: 0.0.1
 homepage: https://github.com/bytedance/flutter_ume
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
