@@ -4,14 +4,14 @@ version: 1.1.0
 homepage: https://github.com/bytedance/flutter_ume
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
     sdk: flutter
   tuple: ^2.0.0
-  share: ^2.0.4
+  share_plus: ^10.0.0
   shared_preferences: ^2.0.6
   flutter_ume: ">=1.0.0 <2.0.0"
 
