@@ -4,16 +4,16 @@ version: 1.0.0
 homepage: https://github.com/bytedance/flutter_ume
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_ume: ">=1.0.0 <2.0.0"
-  system_info: ^1.0.1
-  device_info: ^2.0.2
-  platform: ^3.0.0
+  system_info2: ^4.0.0
+  device_info_plus: ^10.0.0
+  platform: ^3.1.0
 
 dev_dependencies:
   flutter_test:
