@@ -4,8 +4,8 @@ version: 1.1.2
 homepage: https://github.com/bytedance/flutter_ume
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
     sdk: flutter
 
   shared_preferences: ^2.0.15
-  vm_service: ">=9.4.0 <12.0.0"
+  vm_service: ">=14.0.0 <15.0.0"
   tuple: ^2.0.0
 
 dev_dependencies:
